+++ src/main/java/insane96mcp/iguanatweaksreborn/module/mining/Mining.java
@@ -2,6 +2,7 @@ package insane96mcp.iguanatweaksreborn.module.mining;
 
 import insane96mcp.iguanatweaksreborn.module.mining.feature.CustomHardness;
 import insane96mcp.iguanatweaksreborn.module.mining.feature.GlobalHardness;
+import insane96mcp.iguanatweaksreborn.module.mining.feature.InstaMineSilverfish;
 import insane96mcp.iguanatweaksreborn.setup.Config;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -11,12 +12,14 @@ public class Mining extends Module {
 
 	public GlobalHardness globalHardness;
 	public CustomHardness customHardness;
+	public InstaMineSilverfish instaMineSilverfish;
 
 	public Mining() {
 		super(Config.builder);
 		pushConfig(Config.builder);
 		globalHardness = new GlobalHardness(this);
 		customHardness = new CustomHardness(this);
+		instaMineSilverfish = new InstaMineSilverfish(this);
 		Config.builder.pop();
 	}
 
@@ -25,5 +28,6 @@ public class Mining extends Module {
 		super.loadConfig();
 		globalHardness.loadConfig();
 		customHardness.loadConfig();
+		instaMineSilverfish.loadConfig();
 	}
 }